Week seven 1.	Add AdminController.cs and modify AuthController.cs, CommentController.cs, HomeController.cs, MessageController.cs, TagController.cs, PostController.cs, UserController.cs in Controller folder. 2.	Modify ForumDbContext.cs in Data folder. 3.	Modify InitialCreate.Designer.cs, MessageUpdate.cs in Migrations folder. 4.	Modify User.cs, Uses.cs, Comment.cs, Message.cs, Post.cs, Tag.cs and ErrorViewModel.cs in Models folder. 5.	Modify launchSettings.json in the Properties. 6.	Modify SignIn.cshtml and SignUp.cshtml in Auth folder in Views folder. 7.	Modify Index.cshtml and IndexLoggedIn.cshtml in Home folder in Views folder. Add and modify Search.cshtml. 8.	Modify _Layout.cshtml, _layout.cshtml.css, _ValidationScriptsPartial.cshtml and Error.cshtml in the Shared folder. 9.	Add Details.cshtml and modify Create.cshtml, Delete.cshtml, Edit.cshtml and Index.cshtml in Message folder in Views folder. 10.	Modify Create.cshtml, Delete.cshtml, Details.cshtml, Edit.cshtml and Index.cshtml in Post folder in Views folder. 11.	Modify Create.cshtml, Delete.cshtml, Details.cshtml, Edit.cshtml and Index.cshtml in Tag folder in Views folder. 12.	Add Profile.cshtml and modify Create.cshtml, Delete.cshtml, Details.cshtml, Edit.cshtml and Index.cshtml in User folder in Views folder. 13.	Modify _ViewImports.cshtml and _ViewStart.cshtml in View folder. 14.	Modify appsettings.Development.json, appsettings.json, forum.csproj and Program.cs

+++ forum/Controllers/CommentController.cs
@@ -19,99 +19,74 @@ namespace forum.Controllers
             _context = context;
         }
 
+        private async Task<User?> GetUserAsync()
+        {
+            string? token = HttpContext.Session.GetString("token");
+            if (token == null)
+            {
+                return null;
+            }
+            var user = await _context.User.FirstOrDefaultAsync(u => u.token == token);
+            if (user == null)
+            {
+                return null;
+            }
+            return user;
+        }
+
+
         // GET: Comment
         public async Task<IActionResult> Index()
         {
-              return _context.Comment != null ? 
-                          View(await _context.Comment.ToListAsync()) :
-                          Problem("Entity set 'ForumDbContext.Comment'  is null.");
-        }
-
-        // GET: Comment/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null || _context.Comment == null)
+            var user = await GetUserAsync();
+            if (user == null)
             {
-                return NotFound();
+                return RedirectToAction("SignIn", "Auth");
             }
 
-            var comment = await _context.Comment
-                .FirstOrDefaultAsync(m => m.id == id);
-            if (comment == null)
+            string? role = user.role;
+
+            if (role != "admin")
             {
-                return NotFound();
+                return RedirectToAction("SignIn", "Auth");
             }
 
-            return View(comment);
+            return _context.Comment != null ?
+                        View(await _context.Comment.ToListAsync()) :
+                        Problem("Entity set 'ForumDbContext.Comment' is null.");
         }
 
         // GET: Comment/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create(int id)
         {
+            var user = await GetUserAsync();
+            if (user == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+            ViewBag.postId = id;
             return View();
         }
 
         // POST: Comment/Create
-        // To protect from overposting attacks
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("id,userId,postId,content,createdAt")] Comment comment)
+        public async Task<IActionResult> Create([Bind("postId,content")] Comment comment)
         {
+            var user = await GetUserAsync();
+            if (user == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+            var postId = comment.postId;
+            comment.userId = user.id;
             if (ModelState.IsValid)
             {
                 _context.Add(comment);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(comment);
-        }
-
-        // GET: Comment/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null || _context.Comment == null)
-            {
-                return NotFound();
-            }
-
-            var comment = await _context.Comment.FindAsync(id);
-            if (comment == null)
-            {
-                return NotFound();
-            }
-            return View(comment);
-        }
-
-        // POST: Comment/Edit/5
-        // To protect from overposting attacks
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("id,userId,postId,content,createdAt")] Comment comment)
-        {
-            if (id != comment.id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(comment);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!CommentExists(comment.id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("PostComments", "Comment", new { id = postId });
             }
             return View(comment);
         }
@@ -124,11 +99,18 @@ namespace forum.Controllers
                 return NotFound();
             }
 
+            var user = await GetUserAsync();
+
             var comment = await _context.Comment
                 .FirstOrDefaultAsync(m => m.id == id);
             if (comment == null)
             {
                 return NotFound();
+            }
+
+            if (user == null || user.role != "admin" && user.id != comment.userId)
+            {
+                return RedirectToAction("SignIn", "Auth");
             }
 
             return View(comment);
@@ -144,8 +126,15 @@ namespace forum.Controllers
                 return Problem("Entity set 'ForumDbContext.Comment'  is null.");
             }
             var comment = await _context.Comment.FindAsync(id);
+
+            var user = await GetUserAsync();
+
             if (comment != null)
             {
+                if (user == null || user.role != "admin" && user.id != comment.userId)
+                {
+                    return RedirectToAction("SignIn", "Auth");
+                }
                 _context.Comment.Remove(comment);
             }
 
@@ -155,7 +144,22 @@ namespace forum.Controllers
 
         private bool CommentExists(int id)
         {
-          return (_context.Comment?.Any(e => e.id == id)).GetValueOrDefault();
+            return (_context.Comment?.Any(e => e.id == id)).GetValueOrDefault();
+        }
+
+        public async Task<IActionResult> PostComments(int id)
+        {
+            var user = await GetUserAsync();
+            if (user == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
+            var comments = _context.Comment.Where(c => c.postId == id).ToList().OrderByDescending(p => p.createdAt);
+            var post = _context.Post.Where(p => p.id == id).ToList()[0];
+            ViewBag.comments = comments;
+            ViewBag.post = post;
+            return View();
         }
     }
 }
